rename func_body2 module to func_body
