+++ crates/module/src/func_body/instruction.rs
