+++ crates/module/src/func_body/replacer.rs
